new features
ability to stop time
new timer setup

+++ Assets/Scripts/ITimer.cs
@@ -2,8 +2,8 @@
 public interface ITimer
 {
 
-    float GetTime();
+    int GetTime();
 
-    void SetTime(float newTimeLimit);
+    void SetTime(int newTimeLimit);
 
 }
